Update README.md
i want to make the world a better place by bringing together IRL teck to elevate the average persons work week and workload. whether it is buy bringing together scientists and engineers to work on a specific problem or coming up with a billion dollar product to upset the balance of financial power. for example a trading bot that makes small gains or a mining system that reduces the human operator involvement and gives back to the local population allowing for them to improve their living conditions and safety.